Move to a separate file.
