+++ core/o11y/src/lib.rs
@@ -3,6 +3,7 @@
 pub use {backtrace, tracing, tracing_appender, tracing_subscriber};
 
 use clap::Parser;
+pub use context::*;
 use near_crypto::PublicKey;
 use near_primitives::types::AccountId;
 use once_cell::sync::OnceCell;
@@ -15,7 +16,6 @@ use std::borrow::Cow;
 use std::path::PathBuf;
 use tracing::level_filters::LevelFilter;
 use tracing::subscriber::DefaultGuard;
-use tracing::Span;
 use tracing_appender::non_blocking::NonBlocking;
 use tracing_opentelemetry::OpenTelemetryLayer;
 pub use tracing_opentelemetry::OpenTelemetrySpanExt;
@@ -25,6 +25,7 @@ use tracing_subscriber::registry::LookupSpan;
 use tracing_subscriber::{fmt, reload, EnvFilter, Layer, Registry};
 
 /// Custom tracing subscriber implementation that produces IO traces.
+pub mod context;
 mod io_tracer;
 pub mod metrics;
 pub mod testonly;
@@ -582,29 +583,3 @@ macro_rules! log_assert {
         }
     };
 }
-
-/// Wraps an actix message with the current Span's context.
-/// This lets us trace execution across several actix Actors.
-#[derive(actix::Message, Debug)]
-#[rtype(result = "<T as actix::Message>::Result")]
-pub struct WithSpanContext<T: actix::Message> {
-    pub msg: T,
-    pub context: opentelemetry::Context,
-}
-
-impl<T: actix::Message> WithSpanContext<T> {
-    pub fn new(msg: T) -> Self {
-        Self { msg, context: Span::current().context() }
-    }
-}
-
-/// Allows easily attaching the current span's context to a Message.
-pub trait WithSpanContextExt: actix::Message {
-    fn with_span_context(self) -> WithSpanContext<Self>
-    where
-        Self: Sized,
-    {
-        WithSpanContext::<Self>::new(self)
-    }
-}
-impl<T: actix::Message> WithSpanContextExt for T {}
